+++ src/assets/scripts/client/parsers/aircraftCommandParser/aircraftCommandMap.js
@@ -59,6 +59,11 @@ export const AIRCRAFT_COMMAND_MAP = {
         functionName: 'runClimbViaSID',
         isSystemCommand: false
     },
+    cross: {
+        aliases: ['cross', 'cr'],
+        functionName: 'runCross',
+        isSystemCommand: false
+    },
     delete: {
         aliases: ['del', 'delete', 'kill'],
         functionName: 'runDelete',