+++ test/parsers/aircraftCommandParser/argumentValidator.spec.js
@@ -13,7 +13,8 @@ import {
     headingValidator,
     holdValidator,
     squawkValidator,
-    optionalAltitudeValidator
+    optionalAltitudeValidator,
+    crossingValidator
 } from '../../../src/assets/scripts/client/parsers/aircraftCommandParser/argumentValidators';
 
 // TODO: import ERROR_MESSAGE and use actual values to test against
@@ -136,7 +137,7 @@ ava('.altitudeValidator() returns a string when passed anything other than exped
 });
 
 ava('.optionalAltitudeValidator() returns undefined when no value is passed', t => {
-    let result = optionalAltitudeValidator([]);
+    const result = optionalAltitudeValidator([]);
     t.true(typeof result === 'undefined');
 });
 
@@ -294,4 +295,48 @@ ava('.squawkValidator() returns string when passed invalid squawk', t => {
 
     result = squawkValidator(['1a11']);
     t.true(result === 'Invalid argument. Expected \'0000\'-\'7777\' for the transponder code.');
+});
+
+ava('.crossingValidator() returns a string when passed the wrong number of arguments', t => {
+    let result = crossingValidator();
+    t.true(result === 'Invalid argument length. Expected exactly two arguments');
+
+    result = crossingValidator([]);
+    t.true(result === 'Invalid argument length. Expected exactly two arguments');
+
+    result = crossingValidator(['', '', '']);
+    t.true(result === 'Invalid argument length. Expected exactly two arguments');
+});
+
+ava('.crossingValidator() returns undefined when passed valid arguments', t => {
+    let result = crossingValidator(['LEMDY', '50']);
+    t.true(typeof result === 'undefined');
+
+    result = crossingValidator(['BLUB', '100']);
+    t.true(typeof result === 'undefined');
+});
+
+ava('.crossingValidator() returns an error when fixname is not a string', t => {
+    let result = crossingValidator([50, '50']);
+    t.true(result === 'Invalid argument. Must be a string');
+
+    result = crossingValidator([{}, '100']);
+    t.true(result === 'Invalid argument. Must be a string');
+
+    result = crossingValidator([[], '100']);
+    t.true(result === 'Invalid argument. Must be a string');
+});
+
+ava('.crossingValidator() returns an error when altitude is not a number', t => {
+    let result = crossingValidator(['LEMDY', 'xx']);
+    t.true(result === 'Invalid argument. Altitude must be a number');
+
+    result = crossingValidator(['LEMDY', '']);
+    t.true(result === 'Invalid argument. Altitude must be a number');
+
+    result = crossingValidator(['LEMDY', []]);
+    t.true(result === 'Invalid argument. Altitude must be a number');
+
+    result = crossingValidator(['LEMDY', {}]);
+    t.true(result === 'Invalid argument. Altitude must be a number');
 });